combiner.sh updated

+++ combiner.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
-FILES="/home/alex/Downloads/meta/tags/21/*"
-inFile="/home/alex/manyTags.txt"
-rm $inFile
-for f in $FILES
-do
-	echo "Processing $f"
-#	cat $f >> $inFile
-	sed -e :a -e '/$/N; s/\r\n/ /; ta;' $f >> $inFile
-done
+FILES=$1
+inFile=$2
+if [ -d $1 ]; then
+	FILES=$1"*"
+	rm $inFile
+	for f in $FILES
+	do
+		echo "Processing $f"
+		sed -e :a -e '/$/N; s/\r\n/ /; ta;' $f >> $inFile
+	done
+else
+	echo "no dir found"
+fi
